feat: add `LpSlotSet` struct Fixes #27

+++ test/LpToken.t.sol
@@ -3,13 +3,13 @@ pragma solidity ^0.8.17;
 
 import {Test} from "forge-std/Test.sol";
 import {LpToken} from "@usum/core/base/market/LpToken.sol";
-import {IOracleProvider,OracleVersion} from "@usum/core/interfaces/IOracleProvider.sol";
+import {IOracleProvider, OracleVersion} from "@usum/core/interfaces/IOracleProvider.sol";
 import {LpSlotKey, LpSlotKeyLib, Direction} from "@usum/core/libraries/LpSlotKey.sol";
 
 contract OracleProviderMock is IOracleProvider {
-
     // add this to be excluded from coverage report
     function test() public {}
+
     function description() external pure override returns (string memory) {
         return "ETH / USD";
     }
@@ -30,19 +30,15 @@ contract OracleProviderMock is IOracleProvider {
     function pricePrecision() external view override returns (uint256) {}
 }
 
-contract LpTokenTest is Test, LpToken  {
+contract LpTokenTest is Test, LpToken {
     OracleProviderMock public oracleProvider = new OracleProviderMock();
 
     function testUri() public {
-        LpSlotKey id = LpSlotKeyLib.from(uint128(1000), Direction.Long);
-        setImageUri(
-            "https://test.com/images/{id}.png"
-        );
+        LpSlotKey id = LpSlotKeyLib.from(1000, Direction.Long);
+        setImageUri("https://test.com/images/{id}.png");
         // setImageUri(
         //     "https:\\/\\/s3.amazonaws.com\\/your-bucket\\/images\\/{id}.png"
         // );
         emit log_string(uri(id.unwrap()));
     }
-
-
 }
